+++ controllers/contacts/add.js
@@ -1,0 +1,31 @@
+const contactsOperations = require("../../models/contacts");
+
+const Joi = require("joi");
+
+const contactsSchema = Joi.object({
+  name: Joi.string().trim().required(),
+  email: Joi.string().email().trim().required(),
+  phone: Joi.string().required(),
+});
+
+const add = async (req, res, next) => {
+  try {
+    const { error } = contactsSchema.validate(req.body);
+    if (error) {
+      error.status = 400;
+      throw error;
+    }
+    const result = await contactsOperations.addContact(req.body);
+    res.status(201).json({
+      status: "success",
+      code: 201,
+      data: {
+        result,
+      },
+    });
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = add;
