+++ controllers/contacts/getById.js
@@ -1,0 +1,23 @@
+const contactsOperations = require("../../models/contacts");
+const { NotFound } = require("http-errors");
+
+const getById = async (req, res, next) => {
+  try {
+    const { contactId } = req.params;
+    const result = await contactsOperations.getContactById(contactId);
+    if (!result) {
+      throw new NotFound(`Contact with id=${contactId} not found`);
+    }
+    res.json({
+      status: "success",
+      code: 200,
+      data: {
+        result,
+      },
+    });
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = getById;
