+++ controllers/contacts/index.js
@@ -1,0 +1,13 @@
+const getAll = require("./getAll");
+const getById = require("./getById");
+const add = require("./add");
+const removeById = require("./removeById");
+const updateById = require("./updateById");
+
+module.exports = {
+  getAll,
+  getById,
+  add,
+  removeById,
+  updateById,
+};
